added plotting ci and moved plots to examples
